lifted state from recipe editor component
